+++ .vscode/settings.json
@@ -1,5 +1,5 @@
 {
     "yaml.schemas": {
-        "config_schema/config.schema.json" : "config.yaml"
+        "config_schema/config.schema.json": "config.yaml"
     }
 }
