+++ laserstudio/widgets/toolbars/zoomtoolbar.py
@@ -2,7 +2,7 @@ from typing import TYPE_CHECKING
 from PyQt6.QtCore import Qt, QSize
 from PyQt6.QtGui import QIcon
 from PyQt6.QtWidgets import QToolBar, QPushButton
-from ...util import resource_path
+from ...utils.util import resource_path
 
 if TYPE_CHECKING:
     from ...laserstudio import LaserStudio
